extract tech slideshow to own component
